MXS-3403 Handle batched statement in Cache
Earlier the Cache assumed there would be no batched statements,
i.e. a statement arrives before the response to the previous
one has been sent. The implication was that for anybody using
ConnectorJ, the cache would fail in mysterious ways. Not
surprising as the internal logic simply was not prepared for that.

Now, a query that arrives while another query is being handled,
is pushed to a queue. When the complete handling - which may involve
asynchronous communication with the cache storage, the sending
of the query to the backend and the waiting for the response, and
again asynchronous communication with the cache storage - of the
current query is ready, a queued query will (indirectly via the
delayed call mechanism) be popped from the queue and handled.