MXS-2588: Add Consumer class
The Consumer consumes messages from Kafka and orchestrates the batching of
consumed records.

Once either a suitably large batch of records is collected or when an idle
timeout occurs, the Consumer flushes the buffered records into
MariaDB. After this, the offsets of the consumed records are committed
into Kafka itself. This allows a relatively simple method of guaranteeing
that each record is inserted at least once. The downside is that as the
batch size increases, so does the number of records read during
reconnection.

Currently the Consumer also handles the starting and stopping of the
thread it creates to perform all the blocking operations. This doesn't
seem like a very good solution but it is an adequate starting point.