MXS-2588: Improve Consumer error handling
If a fatal error is encountered, the Consumer no longer immediately tries
to continue processing. This will help with short lived problems and
reduce the amount of errors that are logged.

Also extended the info level logging to contain the committed
offsets. This helps detect how fast the records are being processed.