MXS-3417 Default ssl to false
If the value was not set in environment variables, it would
default to true. As MaxScale listeners are not typically configured
for ssl, this would cause issues with some tests that depended on
the value. Defaulting to false is also analogous with backend_ssl.

This change may break some tests that expect the value to be true.
These will be looked at in a later commit.